pyproject.toml
[tool.poetry]
name = "python-db"
version = "0.1.0"
description = ""
authors = ["ytechnologies <mail.ytechnologies.org>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.3"
mysql-connector = "^2.2.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
